Moved memory code from core.hpp to memory.hpp.
