+++ src/pico_tree/pico_tree/cover_tree.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "pico_tree/internal/memory.hpp"
 #include "pico_tree/internal/search_visitor.hpp"
 #include "pico_tree/metric.hpp"
