+++ src/pico_tree/pico_tree/kd_tree.hpp
@@ -3,6 +3,7 @@
 #include <algorithm>
 #include <numeric>
 
+#include "pico_tree/internal/memory.hpp"
 #include "pico_tree/internal/search_visitor.hpp"
 #include "pico_tree/internal/sequence.hpp"
 #include "pico_tree/metric.hpp"
